Comments now have avatar pictures displayed
The pictures are supplied by Gravatar, in case the poster doesn't have a profile there, a identicon is generated from their e-mail hash.
To make it possible to display the avatars right after posting the comment without reloading the page, a new JS library for generating md5 hashes is added.