Completed the "cast" webpage
List of recordings is now presented, just like on the NPC webpage.
Upvoting/downvoting and going to comments works just like on the NPC webpage.
Voice actor's avatar is now also displayed.
A lot of styling is needed, especially in the upper part.
The JS and CSS files for NPC were split and a new set of these files is now present, dedicated specifically for upvoting and downvoting (voting.js and voting.css)