Renamed all NPC avatars so they use the new ID system
All NPCs have now their profile picture associated, but most of them are just default black-and-white profile pictures (but angular). These will be replaced with actual screenshots later.
Some screenshots have their old name, because I couldn't find out which NPC do they belong to. This needs to be fixed by the author of the images in the following commits.