Voting and commenting buttons style change
Change style of the voting and commenting buttons in the lists of recordings of a certain voice actor, or of a certain NPC.
The icons now look more clickable while also having the number of reactions right next to them.
Further styling is probably needed.
A bug with NPC's ID always being set to 0 in the data attributes is fixed too.
Minor JavaScript optimalisations in the voting event handlers.