Complex refactoring of all the controlers.
Controllers are now divided into a few subfolders (and subnamespaces).
New error webpages has been added and the system of error reporting has changed quite a bit.
The process() methods are no longer returning boolean value of true (success) or false (failure, most likely 404), but either 1/true in case of success or the HTTP status code indicating the error (or 204 in case of successful AJAX request).
The errors now have their own namespace and a common parent controller.
This is now also a case for all controllers displaying a website and the one Api controller (who will be split in the next few commits).