Similar refactoring is done with Models
Just as Controllers, Models are now split into subfolders and subnamespaces "Website" and "Api"
The database model is no longer static and it'll be possible to use it for multiple databases with different login details.
For that to be possible, it's now neccessary to pass a link to an INI file containing database credentials to the Db model constructor.