Mod downloads are now improved.
.jar are no longer downloaded directly, but mod downloads go through a download manager that uses the readfile() function.
Because of this, the exception for .jar files has been removed from .htaccess.
A feature allowing to add new downloads will be added in the future (marked with //TODO for now).
The downloads.phtml view will be improved in the future commits and will contain more information about individual downloads.
A new database table containing information about all mod downloads is added.