Add feature allowing admins to create new downloads
In the website administration, there is now a new tab that contains a form allowing admins to create new records in the database "download" table.
This form contains a new TinyMCE textarea.
A new view is added for this form.
A few styling rules are added to administration.css
Two new routes are added to routes.ini, one will be implemented in the future.
The route for downloading .jar files is changed.