Swapping NPC's position is now possible on the npcs page.
This means that administrators can now change the order of NPCs, in which the NPCs are displayed on the mod contents page.
As this feature only added two buttons on the frontend, no styling seems to be required.
Also the table of NPCs in the administration was slightly improved and the NPCs' names are now links just as voice actors' names.