Add system for version checking
The mod client now pings a new route on bootup.
When that happens, the usage analytics are logged, a version is checked and a random fun fact is returned along with the version info.
There is a new config file for the version checker and a library of fun facts (to be expanded later).
The usage analysis controller has been removed and the BootupLogger is now called from a new controller - ModBootupLogger.