System administrators can now manage other users' accounts.
This feature is not meant to be used for changing users' usernames, passwords, bios or profile pictures.
It's meant as a tool for kmaxi to initialize accounts with a profile picture and socials links, if the voice actor doesn't seem like he's willing to fill in the account information and is okay with an admin copying his details from e.g. CCC