Separate downloads page for Forge and Fabric
The /downloads page now serves as a junction, where the user gets to choose between Forge versions and Fabric versions.
After choosing one of the platforms, the user is taken to the downloads list, that is similar to the original downloads page, but contains only downloads for the selected platform.
The junction page will probably need some CSS styling.