Add support for creating releases with remotelly released mod files
Now, direct download links can be specified in the New release form
instead of file name on the local webserver.
When the user chooses to download the JAR file of such download, the JAR
file is downloaded from the 3rd party server to ours and then forwarded
to the client in real time.